Spring Data JPA

+++ src/main/java/ru/job4j/social/media/api/repository/FriendRepository.java
@@ -4,5 +4,5 @@ package ru.job4j.social.media.api.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.social.media.api.model.Friend;
 
-public interface FriendRepository extends CrudRepository<Friend, Integer> {
+public interface FriendRepository extends CrudRepository<Friend, Long> {
 }
